refactor: change architecture ws_servers
