+++ src/common/constants.ts
@@ -1,0 +1,2 @@
+export const HTTP_PORT = 3000;
+export const WS_PORT = 8080;
